refactor dtos and fix pom

+++ src/main/java/com/lyahov/stepa/MyTodoIstBack/business/entity/ProjectEntity.java
@@ -18,7 +18,8 @@ import java.util.List;
 public class ProjectEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    Integer id;
+    Long id;
+
     String title;
 
     @OneToMany(mappedBy = "projectId", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, fetch = FetchType.EAGER)
